requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
alembic==1.16.4
    # via dagster
annotated-types==0.7.0
    # via pydantic
anthropic==0.64.0
    # via dagster-dg-cli
antlr4-python3-runtime==4.13.2
    # via dagster
anyio==4.10.0
    # via
    #   anthropic
    #   claude-code-sdk
    #   gql
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
    #   watchfiles
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
backoff==2.2.1
    # via gql
beautifulsoup4==4.13.4
    # via yfinance
certifi==2025.8.3
    # via
    #   curl-cffi
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via
    #   cryptography
    #   curl-cffi
charset-normalizer==3.4.3
    # via requests
claude-code-sdk==0.0.20
    # via dagster-dg-cli
click==8.1.8
    # via
    #   click-aliases
    #   dagster
    #   dagster-cloud-cli
    #   dagster-dg-core
    #   dagster-webserver
    #   typer
    #   uvicorn
click-aliases==1.0.5
    # via dagster-dg-core
coloredlogs==14.0
    # via dagster
cryptography==45.0.6
    # via pyjwt
curl-cffi==0.13.0
    # via yfinance
dagster==1.11.7
    # via
    #   stock-data-pipeline (pyproject.toml)
    #   dagster-dg-cli
    #   dagster-graphql
    #   dagster-webserver
dagster-cloud-cli==1.11.7
    # via dagster-dg-core
dagster-dg-cli==1.11.7
    # via stock-data-pipeline (pyproject.toml)
dagster-dg-core==1.11.7
    # via dagster-dg-cli
dagster-graphql==1.11.7
    # via dagster-webserver
dagster-pipes==1.11.7
    # via dagster
dagster-shared==1.11.7
    # via
    #   dagster
    #   dagster-cloud-cli
    #   dagster-dg-core
dagster-webserver==1.11.7
    # via stock-data-pipeline (pyproject.toml)
distro==1.9.0
    # via anthropic
docstring-parser==0.17.0
    # via dagster
dotenv==0.9.9
    # via stock-data-pipeline (pyproject.toml)
filelock==3.19.1
    # via dagster
frozendict==2.4.6
    # via yfinance
fsspec==2025.7.0
    # via universal-pathlib
github3-py==4.0.1
    # via dagster-cloud-cli
gql==3.5.3
    # via
    #   dagster-dg-core
    #   dagster-graphql
graphene==3.4.3
    # via dagster-graphql
graphql-core==3.2.6
    # via
    #   gql
    #   graphene
    #   graphql-relay
graphql-relay==3.2.0
    # via graphene
greenlet==3.2.4
    # via sqlalchemy
grpcio==1.74.0
    # via
    #   dagster
    #   grpcio-health-checking
grpcio-health-checking==1.71.2
    # via dagster
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   anthropic
    #   mcp
httpx-sse==0.4.1
    # via mcp
humanfriendly==10.0
    # via coloredlogs
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.6
    # via
    #   dagster
    #   dagster-cloud-cli
    #   dagster-dg-core
jiter==0.10.0
    # via anthropic
jsonschema==4.25.1
    # via
    #   dagster-dg-core
    #   mcp
jsonschema-specifications==2025.4.1
    # via jsonschema
mako==1.3.10
    # via alembic
markdown==3.8.2
    # via dagster-dg-core
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
mcp==1.13.1
    # via dagster-dg-cli
mdurl==0.1.2
    # via markdown-it-py
multidict==6.6.4
    # via yarl
multitasking==0.0.12
    # via yfinance
numpy==2.3.2
    # via
    #   pandas
    #   yfinance
packaging==25.0
    # via
    #   dagster-cloud-cli
    #   dagster-dg-core
    #   dagster-shared
pandas==2.3.2
    # via yfinance
peewee==3.18.2
    # via yfinance
platformdirs==4.3.8
    # via yfinance
prompt-toolkit==3.0.51
    # via questionary
propcache==0.3.2
    # via yarl
protobuf==5.29.5
    # via
    #   dagster
    #   grpcio-health-checking
    #   yfinance
psycopg2-binary==2.9.10
    # via stock-data-pipeline (pyproject.toml)
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   anthropic
    #   dagster-shared
    #   mcp
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via mcp
pygments==2.19.2
    # via rich
pyjwt==2.10.1
    # via github3-py
python-dateutil==2.9.0.post0
    # via
    #   github3-py
    #   graphene
    #   pandas
python-dotenv==1.1.1
    # via
    #   dagster
    #   dagster-dg-core
    #   dotenv
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via mcp
pytz==2025.2
    # via
    #   dagster
    #   pandas
    #   yfinance
pyyaml==6.0.2
    # via
    #   dagster-cloud-cli
    #   dagster-dg-core
    #   dagster-shared
    #   uvicorn
questionary==2.1.0
    # via dagster-cloud-cli
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   stock-data-pipeline (pyproject.toml)
    #   dagster
    #   dagster-cloud-cli
    #   dagster-graphql
    #   github3-py
    #   gql
    #   requests-toolbelt
    #   yfinance
requests-toolbelt==1.0.0
    # via gql
rich==14.1.0
    # via
    #   dagster
    #   dagster-dg-core
    #   typer
rpds-py==0.27.0
    # via
    #   jsonschema
    #   referencing
setuptools==80.9.0
    # via
    #   dagster
    #   dagster-cloud-cli
    #   dagster-dg-core
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   dagster
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
soupsieve==2.7
    # via beautifulsoup4
sqlalchemy==2.0.43
    # via
    #   alembic
    #   dagster
sse-starlette==3.0.2
    # via mcp
starlette==0.47.2
    # via
    #   dagster-graphql
    #   dagster-webserver
    #   mcp
structlog==25.4.0
    # via dagster
tabulate==0.9.0
    # via dagster
termcolor==2.3.0
    # via yaspin
tomli==2.2.1
    # via dagster
tomlkit==0.13.2
    # via
    #   dagster-dg-core
    #   dagster-shared
toposort==1.10
    # via dagster
tqdm==4.67.1
    # via dagster
typer==0.16.1
    # via
    #   dagster-cloud-cli
    #   dagster-dg-cli
    #   dagster-dg-core
typing-extensions==4.14.1
    # via
    #   alembic
    #   anthropic
    #   beautifulsoup4
    #   dagster-dg-core
    #   dagster-shared
    #   graphene
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via
    #   dagster
    #   pandas
universal-pathlib==0.2.6
    # via dagster
uritemplate==4.2.0
    # via github3-py
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via
    #   dagster-webserver
    #   mcp
uvloop==0.21.0
    # via uvicorn
watchdog==5.0.3
    # via
    #   dagster
    #   dagster-dg-core
watchfiles==1.1.0
    # via uvicorn
wcwidth==0.2.13
    # via prompt-toolkit
websockets==15.0.1
    # via
    #   uvicorn
    #   yfinance
yarl==1.20.1
    # via gql
yaspin==3.1.0
    # via dagster-dg-core
yfinance==0.2.65
    # via stock-data-pipeline (pyproject.toml)
